// Project 1/ClientApp/src/app/blocks/enum/file-system-tags.enum.ts
export enum FileSystemTagsEnum {
    generalSpeciality = 'general',
    cardiologySpeciality = 'cardiology',
    condition = 'condition',
    followup = 'followup',
    operation = 'operation',
    laboratory = 'laboratory',
    physicalExam = 'physicalExam',
    radio = 'radio',
    ticket='ticket'

}